coldata: make coldata.BatchSize a package function
This will allow us to vary the batchSize private variable int the future
by providing a testing hook for external packages to modify the batch
size at will.

Release justification: Category 1 non-production code change.

Release note: None

+++ pkg/sql/colexec/execpb/stats.go
@@ -40,7 +40,7 @@ func (vs *VectorizedStats) Stats() map[string]string {
 	}
 	selectivity := float64(0)
 	if vs.NumBatches > 0 {
-		selectivity = float64(vs.NumTuples) / float64(coldata.BatchSize*vs.NumBatches)
+		selectivity = float64(vs.NumTuples) / float64(int64(coldata.BatchSize())*vs.NumBatches)
 	}
 	return map[string]string{
 		batchesOutputTagSuffix: fmt.Sprintf("%d", vs.NumBatches),
@@ -68,7 +68,7 @@ func (vs *VectorizedStats) StatsForQueryPlan() []string {
 	}
 	selectivity := float64(0)
 	if vs.NumBatches > 0 {
-		selectivity = float64(vs.NumTuples) / float64(coldata.BatchSize*vs.NumBatches)
+		selectivity = float64(vs.NumTuples) / float64(int64(coldata.BatchSize())*vs.NumBatches)
 	}
 	return []string{
 		fmt.Sprintf("%s: %d", batchesOutputQueryPlanSuffix, vs.NumBatches),
